build(deps)!: Upgrade to opentelemetry 0.28
This is required to upgrade to make log correlation continue to work, as
otherwise the `OTelData` struct will be referencing the wrong crate
version, causing span extension extraction to fail. I've confirmed this
works with our service that is on the latest versions of opentelemetry &
tracing.

It's somewhat chunky because opentelemetry dropped the requirement for
an async runtime, which means reporting is done through the reqwest
blocking client. Due to some peculiarities of how that works, the
opentelemetry instrumentation cannot be setup in an async context
anymore, lest one gets "cannot drop runtime" errors.

I've also adjusted the test cases and documentation to the new APIs, and
made sure all tests pass in the current form.

There are still some deprecation warnings in the otel example concerning
the trace config builder, but I deemed those non-essential for the time
being.