build(deps)!: Upgrade to opentelemetry 0.28
This is required to upgrade to make log correlation continue to work, as
otherwise the `OTelData` struct will be referencing the wrong crate
version, causing span extension extraction to fail.

It's somewhat chunky because opentelemetry dropped the requirement for
an async runtime, which means reporting is done through the reqwest
blocking client. Due to some peculiarities of how that works, the
opentelemetry instrumentation cannot be setup in an async context
anymore, lest one gets "cannot drop runtime" errors.

BREAKING CHANGE: upgrading the otel version is deemed as breaking as the otel code might break downstream code silently